refactor: move ModelEditor component to model-editor directory and update imports
